+++ .idea/modules.xml
@@ -3,7 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
-      <module fileurl="file://$PROJECT_DIR$/googlecards.iml" filepath="$PROJECT_DIR$/googlecards.iml" />
+      <module fileurl="file://$PROJECT_DIR$/googlecards2.iml" filepath="$PROJECT_DIR$/googlecards2.iml" />
     </modules>
   </component>
 </project>
