+++ app/src/main/AndroidManifest.xml
@@ -40,7 +40,7 @@
         <activity
             android:name=".MainActivity"
             android:label="@string/app_name"
-            android:screenOrientation="portrait"
+            android:screenOrientation="landscape"
             android:windowSoftInputMode="stateHidden" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -51,8 +51,13 @@
         <activity
             android:name=".ShareNote"
             android:label="@string/title_activity_share_note" >
+
+        </activity>
+
+        <activity
             android:name=".CalendarQuickstart"
-            android:label="@string/title_activity_calendar_quickstart" >
+            android:label="@string/title_activity_calendar_quickstart">
+
         </activity>
 
         <meta-data
