+++ app/src/main/res/layout/fragment_left.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="cards"
+        android:textStyle="bold"
+        android:id="@+id/cardsTitle"
+        android:gravity="start"
+        android:paddingLeft="10dp"
+        android:textSize="40sp"
+        android:textColor="#2196f3"/>
+
+   <ListView
+       android:layout_width="match_parent"
+       android:layout_height="match_parent"
+       android:id="@+id/listView"
+       android:layout_below="@+id/cardsTitle"/>
+
+</RelativeLayout>
